Divide UI logic from MainActivity
