+++ app/src/main/java/com/example/androiddevchallenge/MainActivity.kt
@@ -18,44 +18,17 @@ package com.example.androiddevchallenge
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredWidth
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Scaffold
 import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navArgument
-import androidx.navigation.compose.navigate
 import androidx.navigation.compose.rememberNavController
-import com.example.androiddevchallenge.model.Cat
-import com.example.androiddevchallenge.model.catsSeed
+import com.example.androiddevchallenge.ui.screens.CatDetailScreen
+import com.example.androiddevchallenge.ui.screens.CatListScreen
 import com.example.androiddevchallenge.ui.theme.MyTheme
 
 class MainActivity : AppCompatActivity() {
@@ -101,86 +74,15 @@ fun NavGraph(startDestination: String = "cats") {
         navController = navController,
         startDestination = startDestination
     ) {
-        composable("cats") { CatList(navController) }
+        composable("cats") { CatListScreen(navController) }
         composable(
             "cat/{catId}",
             arguments = listOf(navArgument("catId") { type = NavType.IntType })
         ) { backStackEntry ->
-            CatDetail(
+            CatDetailScreen(
                 navController,
                 requireNotNull(backStackEntry.arguments).getInt("catId")
             )
-        }
-    }
-}
-
-@Composable
-fun CatList(
-    navController: NavController,
-    cats: List<Cat> = catsSeed
-) {
-    LazyColumn(Modifier.fillMaxSize()) {
-        items(cats) { cat ->
-            CatCard(cat) { navController.navigate("cat/${cat.id}") }
-        }
-    }
-}
-
-@Composable
-fun CatCard(
-    cat: Cat,
-    onClick: () -> Unit
-) {
-    Row(
-        Modifier
-            .padding(8.dp)
-            .clip(RoundedCornerShape(4.dp))
-            .background(MaterialTheme.colors.surface)
-            .clickable(onClick = onClick)
-            .fillMaxWidth()
-    ) {
-        Surface(
-            modifier = Modifier.size(80.dp),
-            shape = CircleShape,
-            color = MaterialTheme.colors.onSurface.copy(alpha = 0.2f)
-        ) {
-            Image(painter = painterResource(cat.image), contentDescription = cat.name)
-        }
-        Spacer(Modifier.requiredWidth(24.dp))
-        Text(cat.name, fontSize = 24.sp)
-    }
-}
-
-@Composable
-fun CatDetail(
-    navController: NavController,
-    catId: Int
-) {
-    catsSeed.find { it.id == catId }?.also { cat ->
-        CatDetailView(cat)
-    } ?: run {
-        navController.navigateUp()
-    }
-}
-
-@Composable
-fun CatDetailView(cat: Cat) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = {
-                    Text(cat.name)
-                }
-            )
-        }
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Image(painter = painterResource(cat.image), contentDescription = cat.name)
-            Text(text = cat.name)
         }
     }
 }
