+++ app/src/main/java/com/example/androiddevchallenge/ui/screens/CatDetailScreen.kt
@@ -1,0 +1,65 @@
+/*
+ * Copyright 2021 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.example.androiddevchallenge.ui.screens
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.navigation.NavController
+import com.example.androiddevchallenge.model.Cat
+import com.example.androiddevchallenge.model.catsSeed
+
+@Composable
+fun CatDetailScreen(
+    navController: NavController,
+    catId: Int
+) {
+    catsSeed.find { it.id == catId }?.also { cat ->
+        CatDetailView(cat)
+    } ?: run {
+        navController.navigateUp()
+    }
+}
+
+@Composable
+fun CatDetailView(cat: Cat) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(cat.name)
+                }
+            )
+        }
+    ) {
+        Column(
+            modifier = Modifier.fillMaxWidth(),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Image(painter = painterResource(cat.image), contentDescription = cat.name)
+            Text(text = cat.name)
+        }
+    }
+}
